Add separated date range picker component
