+++ src/DateRangePickerMenu/index.tsx
@@ -7,88 +7,41 @@
  *
  * SPDX-License-Identifier: GPL-3.0-only
  ********************************************************************************/
-import { CalendarTwoTone } from '@ant-design/icons';
-import { DatePicker, Divider, Space, Typography } from 'antd';
-import moment, { Duration, Moment } from 'moment';
-import React, { useEffect, useRef, useState } from 'react';
-import styles from './DateRangePickerMenu.module.scss';
 
-interface DateRangePickerMenuProps {
-  dateFormat?: string;
-  ranges?: Record<string, Duration>;
-  onChange(dates: [Moment, Moment]): void;
+import { G2 } from '@ant-design/charts';
+import { Datum, Meta } from '@antv/g2plot/lib/types';
+import { partition, pick } from 'lodash-es';
+import moment, { Moment } from 'moment';
+import React from 'react';
+import DateRangePickerMenu from './DateRangePickerMenu';
+
+function getXYScales(scales: Record<string, Meta>) {
+  const isTimeAxis = (axis: string) => scales[axis].type === 'timeCat' || scales[axis].type === 'time';
+  const [xScaleNames, yScaleNames] = partition(Object.keys(scales), isTimeAxis);
+  const xScales = pick(scales, xScaleNames);
+  const yScales = pick(scales, yScaleNames);
+  return { xScales, yScales };
 }
 
-const defaultRanges: Record<string, Duration> = {
-  '1D': moment.duration(1, 'd'),
-  '1W': moment.duration(1, 'w'),
-  '1M': moment.duration(1, 'M'),
-  '3M': moment.duration(3, 'M'),
-  '6M': moment.duration(6, 'M'),
-  '1Y': moment.duration(1, 'y'),
+const ChartDateRangePicker = ({ plot, options }: any) => {
+  const handleDatesChanged = (dates: [Moment, Moment]) => {
+    const { timeUnit = null } = options;
+    const { chart } = plot;
+    const chartViews = [...(chart.views || []), ...(chart.options.scales ? [chart] : [])];
+    chartViews.forEach((view: any) => {
+      const scales = view.options.scales || chart.options.scales;
+      const { xScales } = getXYScales(scales);
+      const timeScalesName = Object.keys(xScales)[0];
+      view.filter('*', (value: any, data: Datum) => {
+        const resultDate = moment(data[timeScalesName]);
+        return resultDate.isSameOrAfter(dates[0], timeUnit) && resultDate.isSameOrBefore(dates[1], timeUnit);
+      });
+    });
+    chart.render();
+    chart.emit('data:filter', G2.Event.fromData(plot?.chart, 'data:filter', { timestamp: Date.now() }));
+  };
+
+  return <DateRangePickerMenu onChange={handleDatesChanged} />;
 };
 
-const DateRangePickerMenu: React.FC<DateRangePickerMenuProps> = ({
-  dateFormat = 'DD.MM.YYYY',
-  ranges = defaultRanges,
-  onChange,
-}) => {
-  const [dates, setDates] = useState<[Moment, Moment] | null>(null);
-  const activeRangeRef = useRef<any>(null);
-  const { RangePicker } = DatePicker;
-
-  useEffect(() => {
-    if (dates) {
-      onChange(dates);
-    }
-  }, [dates]);
-
-  return (
-    <div className={styles.controlPanel}>
-      <div>
-        <Space split={<Divider type='vertical' style={{ borderWidth: 2, margin: 0, borderColor: '#dadce0' }} />}>
-          {Object.keys(ranges).map((range: string) => (
-            <Typography.Link
-              key={range}
-              onClick={(e) => {
-                const calendarDates: [Moment, Moment] = [moment().subtract(ranges[range]), moment()];
-                if (activeRangeRef.current !== null) {
-                  activeRangeRef.current.classList.remove(styles.activeRange);
-                }
-                activeRangeRef.current = e.target;
-                activeRangeRef.current.classList.add(styles.activeRange);
-                setDates(calendarDates);
-              }}>
-              {range}
-            </Typography.Link>
-          ))}
-        </Space>
-      </div>
-
-      <div className={styles.rangePicker}>
-        <RangePicker
-          inputReadOnly
-          allowClear={false}
-          bordered={false}
-          showNow
-          separator='-'
-          onChange={(calendarDates: any) => {
-            if (activeRangeRef.current !== null) {
-              activeRangeRef.current.classList.remove(styles.activeRange);
-            }
-            setDates(calendarDates);
-            onChange(calendarDates);
-          }}
-          disabledDate={(d) => !d || d.isAfter(moment())}
-          className={styles.picker}
-          size='small'
-          value={dates}
-          format={dateFormat}
-          suffixIcon={<CalendarTwoTone className={styles.suffix} />}
-        />
-      </div>
-    </div>
-  );
-};
-
-export default DateRangePickerMenu;
+export default ChartDateRangePicker;
